gateway-service/requirements.txt
attrs==21.4.0
boto3==1.20.46
botocore==1.23.46
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
cycler==0.11.0
Flask==2.0.2
Flask-Cors==3.0.10
fonttools==4.29.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
kafka-python==2.0.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
nexradaws==1.1
numpy==1.22.1
packaging==21.3
Pillow==9.0.0
pluggy==1.0.0
py==1.11.0
pymongo==4.0.1
pyparsing==3.0.7
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
s3transfer==0.5.0
six==1.16.0
toml==0.10.2
urllib3==1.26.8
Werkzeug==2.0.2
